[mlir] Fix a warning
This patch fixes:

  mlir/lib/Dialect/SparseTensor/IR/SparseTensorDialect.cpp:675:56:
  error: comparison of integers of different signs: 'long const' and
  'const mlir::sparse_tensor::Level' (aka 'const unsigned long')
  [-Werror,-Wsign-compare]

+++ mlir/lib/Dialect/SparseTensor/IR/SparseTensorDialect.cpp
@@ -672,7 +672,7 @@ static LogicalResult verifyPackUnPack(Operation *op, bool requiresStaticShape,
   // level-coordinates (cf., the op documentation).
   const auto coordsRank = coordinatesTp.getShape()[1];
   const auto tensorRank = tensorTp.getLvlRank();
-  if (!ShapedType::isDynamic(coordsRank) && coordsRank != tensorRank)
+  if (!ShapedType::isDynamic(coordsRank) && (unsigned)coordsRank != tensorRank)
     return op->emitError("input/output level-ranks don't match");
 
   return success();
